backend/requirements.txt
# Streaming Backend — Python dependencies
# Install: pip install -r requirements.txt
#
# System requirement (not pip-installable):
#   FFmpeg — required for merging video chunks.
#   macOS:   brew install ffmpeg
#   Ubuntu:  sudo apt install ffmpeg
#   Windows: https://ffmpeg.org/download.html

# --- Core (required) ---
fastapi==0.115.6
websockets>=12.0
uvicorn[standard]==0.32.1
pydantic>=2.12.5
python-multipart==0.0.17
python-dotenv==1.0.1
httpx==0.27.0
edge-tts==6.1.18
websockets>=12.0

# --- AI (resume-ats, skills, JD – dynamic config from Superadmin) ---
openai>=1.0.0

# --- Proctoring (face detection + head pose in backend) ---
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# --- STT: AssemblyAI streaming (same as KareerGrowth AiService; SDK avoids token 422) ---
assemblyai>=0.34.0
scipy>=1.11.0
certifi>=2024.0.0

# --- Background tasks (optional; merge runs in-process if not installed) ---
celery[redis]==5.4.0
redis==5.2.1

# --- MongoDB async driver (used by report_generator) ---
motor>=3.3.0

# --- Cloud storage (optional; uncomment if using) ---
# boto3==1.35.93
# google-cloud-storage==2.18.4
